pyproject.toml
[tool.poetry]
name = "resources_api"
version = "0.1.0"
description = "Resources API for Operation Code"
authors = ["OC_Contributors"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
algoliasearch = ">=2.0,<3.0"
alembic = "1.4.2"
bandit = "1.5.1"
click = "7.1.2"
flake8 = "3.8.3"
flask = "1.1.2"
Flask-Cors = "3.0.8"
Flask-Limiter = "1.3.1"
Flask-Migrate = "2.5.3"
prometheus_client = "0.8.0"
psycopg2-binary = "2.8.5"
py-healthcheck = "1.10.1"
pytest = "5.4.3"
pytest-mock = "3.1.1"
pytest-cov = "2.10.0"
python-dateutil = "2.8.1"
pyyaml = "5.3.1"
requests = "2.23.0"
sqlalchemy = "1.3.17"
SQLAlchemy-Utils = "0.36.6"
uWSGI = "2.0.18"
Werkzeug = "1.0.1"
pyjwt = "^1.7.1"
cryptography = "^2.9.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
